% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSECDaily.R
\name{downloadSECDaily}
\alias{downloadSECDaily}
\title{Download an index for a certain day.}
\usage{
downloadSECDaily(date)
}
\arguments{
\item{date}{a `Date` type object for the date to download.}
}
\value{
returns a data.table.
}
\description{
\code{downloadSECQtr} Download  an index for a certain day.
}
\details{
Downloads an SEC index file for a day. This function uses the
downloadSECIndex function and thus needs the `USERAGENT` environment
variable set.
}
\examples{
\dontrun{

link <- "https://www.sec.gov/Archives/edgar/daily-index/2024/QTR4/master.20241004.idx"
output <- downloadSECDaily("2024-10-04")
output
}
}
